+++ src/pymonzo/exceptions.py
@@ -10,16 +10,11 @@ class PyMonzoException(Exception):
     pass
 
 
-class MonzoAPIException(PyMonzoException):
+class MonzoAPIError(PyMonzoException):
     """Monzo API response exception"""
     pass
 
 
-class UnableToGetToken(PyMonzoException):
-    """Unable to get a token generic"""
-    pass
-
-
-class UnableToRefreshTokenException(MonzoAPIException):
+class CantRefreshTokenError(MonzoAPIError):
     """Base Monzo API response exception"""
     pass