+++ CHANGELOG.md
@@ -10,6 +10,7 @@ adheres to [Semantic Versioning][semver].
   current token and raises `UnableToRefreshTokenException` when token couldn't
   be refreshed.
 - Client secret is now saved in token file JSON file.
+- Cleaned up exceptions.
 
 ## [v0.10.1][v0.10.1] - 2017-09-24
 ### Fixed